Broke up file structure to proper flask with templates and static directories. Also added new field to tables
