+++ requirements.txt
@@ -2,3 +2,4 @@ Flask==2.3.3
 Flask-SQLAlchemy==3.0.5
 SQLAlchemy==2.0.21
 Werkzeug==2.3.7
+Flask-Migrate==4.0.4
